# WebApiTransJ/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebApiTransJ, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Bitacora/RegistrarBitacora": {
      "post": {
        "tags": [
          "Bitacora"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BitacoraViajeEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Bitacora/ActualizarBitacora": {
      "put": {
        "tags": [
          "Bitacora"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BitacoraViajeEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Bitacora/CambiarEstadoBitacora": {
      "put": {
        "tags": [
          "Bitacora"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdBitacora",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "IdUsuario",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Bitacora/ListarBitacoraViajeXID": {
      "get": {
        "tags": [
          "Bitacora"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdBItacora",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Catalogo/ListarMunicipio": {
      "get": {
        "tags": [
          "Catalogo"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "departamento",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Catalogo/UsuarioOrganizacion": {
      "get": {
        "tags": [
          "Catalogo"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdUsuario",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Catalogo/UsuarioRoles": {
      "get": {
        "tags": [
          "Catalogo"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdUsuario",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Horarios/RegistrarHorario": {
      "post": {
        "tags": [
          "Horarios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/HorarioEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Horarios/ActualizarHorario": {
      "put": {
        "tags": [
          "Horarios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/HorarioEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Horarios/CambiarEstadoHorario": {
      "put": {
        "tags": [
          "Horarios"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdHorario",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "IdUsuario",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Horarios/listarHorariosXID": {
      "get": {
        "tags": [
          "Horarios"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdHorario",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "BitacoraViajeEntity": {
      "type": "object",
      "properties": {
        "pIdNumeroEntrega": {
          "type": "string"
        },
        "pFechaEntrega": {
          "type": "string"
        },
        "pHoraEntrega": {
          "type": "string"
        },
        "pDiaEntrega": {
          "type": "string"
        },
        "pLugarEntrega": {
          "type": "string"
        },
        "pProductoEntrega": {
          "type": "string"
        },
        "pIdUsuario": {
          "type": "string"
        },
        "pObservaciones": {
          "type": "string"
        },
        "pIdPiloto": {
          "format": "int32",
          "type": "integer"
        },
        "pIdUnidad": {
          "format": "int32",
          "type": "integer"
        },
        "pIdSocio": {
          "format": "int32",
          "type": "integer"
        },
        "pIdDepartamento": {
          "format": "int32",
          "type": "integer"
        },
        "pIdMunicipio": {
          "format": "int32",
          "type": "integer"
        },
        "pFechaCreacion": {
          "type": "string"
        },
        "pTransaccionEstado": {
          "format": "int32",
          "type": "integer"
        },
        "pTransaccionMensaje": {
          "type": "string"
        },
        "pIdBitacora": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "HorarioEntity": {
      "type": "object",
      "properties": {
        "pIdHorario": {
          "format": "int32",
          "type": "integer"
        },
        "pHoraEntrega": {
          "type": "string"
        },
        "pFechaCarga": {
          "type": "string"
        },
        "pIdUsuario": {
          "type": "string"
        },
        "pPlanta": {
          "type": "string"
        },
        "pIdPiloto": {
          "format": "int32",
          "type": "integer"
        },
        "pIdUnidad": {
          "format": "int32",
          "type": "integer"
        },
        "pIdSocio": {
          "format": "int32",
          "type": "integer"
        },
        "pTransaccionEstado": {
          "format": "int32",
          "type": "integer"
        },
        "pTransaccionMensaje": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  }
}